Moved all handler functionality into ts.handler_utils
