+++ ts/torch_handler/image_classifier.py
@@ -1,12 +1,11 @@
 """
 Module for image classification default handler
 """
-import torch
-import torch.nn.functional as F
 from torchvision import transforms
 
+from ts.handler_utils import ImageClassificationPostprocess, VisionPreproc
+
 from .vision_handler import VisionHandler
-from ..utils.util  import map_class_to_label
 
 
 class ImageClassifier(VisionHandler):
@@ -15,26 +14,18 @@ class ImageClassifier(VisionHandler):
     and returns the name of object in that image.
     """
 
-    topk = 5
     # These are the standard Imagenet dimensions
     # and statistics
-    image_processing = transforms.Compose([
-        transforms.Resize(256),
-        transforms.CenterCrop(224),
-        transforms.ToTensor(),
-        transforms.Normalize(mean=[0.485, 0.456, 0.406],
-                             std=[0.229, 0.224, 0.225])
-    ])
+    image_processing = transforms.Compose(
+        [
+            transforms.Resize(256),
+            transforms.CenterCrop(224),
+            transforms.ToTensor(),
+            transforms.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225]),
+        ]
+    )
 
-    def set_max_result_classes(self, topk):
-        self.topk = topk
-
-    def get_max_result_classes(self):
-        return self.topk
-
-    def postprocess(self, data):
-        ps = F.softmax(data, dim=1)
-        probs, classes = torch.topk(ps, self.topk, dim=1)
-        probs = probs.tolist()
-        classes = classes.tolist()
-        return map_class_to_label(probs, self.mapping, classes)
+    def __init__(self):
+        super().__init__()
+        self.preprocess = VisionPreproc()
+        self.postprocess = ImageClassificationPostprocess()
