+++ ts/torch_handler/image_segmenter.py
@@ -1,9 +1,12 @@
 """
 Module for image segmentation default handler
 """
-import torch
 from torchvision import transforms as T
+
+from ts.handler_utils import ImageSegmentationPostprocess, VisionPreproc
+
 from .vision_handler import VisionHandler
+
 
 class ImageSegmenter(VisionHandler):
     """
@@ -12,21 +15,16 @@ class ImageSegmenter(VisionHandler):
     where N - batch size, K - number of classes, H - height and W - width.
     """
 
-    image_processing = T.Compose([
-        T.Resize(256),
-        T.CenterCrop(224),
-        T.ToTensor(),
-        T.Normalize(
-            mean=[0.485, 0.456, 0.406],
-            std=[0.229, 0.224, 0.225]
-        )])
+    image_processing = T.Compose(
+        [
+            T.Resize(256),
+            T.CenterCrop(224),
+            T.ToTensor(),
+            T.Normalize(mean=[0.485, 0.456, 0.406], std=[0.229, 0.224, 0.225]),
+        ]
+    )
 
-    def postprocess(self, data):
-        # Returning the class for every pixel makes the response size too big
-        # (> 24mb). Instead, we'll only return the top class for each image
-        data = data['out']
-        data = torch.nn.functional.softmax(data, dim=1)
-        data = torch.max(data, dim=1)
-        data = torch.stack([data.indices.type(data.values.dtype), data.values], dim=3)
-
-        return data.tolist()
+    def __init__(self):
+        super().__init__()
+        self.preprocess = VisionPreproc()
+        self.postprocess = ImageSegmentationPostprocess()
