+++ ts/torch_handler/vision_handler.py
@@ -3,18 +3,13 @@
 """
 Base module for all vision handlers
 """
-import base64
-import io
-import types
 from abc import ABC
 
-import torch
-from PIL import Image
+from torchvision import transforms
+
+from ts.handler_utils import BaseInit, VisionInit, VisionPreproc
 
 from .base_handler import BaseHandler
-
-print(__file__)
-from ts.torch_handler.handler_utils import BaseInit, VisionInit
 
 
 class VisionHandler(BaseHandler, ABC):
@@ -22,40 +17,12 @@ class VisionHandler(BaseHandler, ABC):
     Base class for all vision handlers
     """
 
+    image_processing = transforms.Compose([transforms.ToTensor()])
+
     def __init__(self):
-
-        self.initialize = types.MethodType(VisionInit(BaseInit()), self)
-
-    def preprocess(self, data):
-        """The preprocess function of MNIST program converts the input data to a float tensor
-
-        Args:
-            data (List): Input data from the request is in the form of a Tensor
-
-        Returns:
-            list : The preprocess function returns the input image as a list of float tensors.
-        """
-        images = []
-
-        for row in data:
-            # Compat layer: normally the envelope should just return the data
-            # directly, but older versions of Torchserve didn't have envelope.
-            image = row.get("data") or row.get("body")
-            if isinstance(image, str):
-                # if the image is a string of bytesarray.
-                image = base64.b64decode(image)
-
-            # If the image is sent as bytesarray
-            if isinstance(image, (bytearray, bytes)):
-                image = Image.open(io.BytesIO(image))
-                image = self.image_processing(image)
-            else:
-                # if the image is a list
-                image = torch.FloatTensor(image)
-
-            images.append(image)
-
-        return torch.stack(images).to(self.device)
+        super().__init__()
+        self.initialize = VisionInit(BaseInit())
+        self.preprocess = VisionPreproc()
 
     def get_insights(self, tensor_data, _, target=0):
         print("input shape", tensor_data.shape)
